Fix casting type of DishContent
- Change VariantName color

+++ src/ARMenu/Assets/Scripts/CameraScreenScripts/DetailsControl/DetailsControl.cs
@@ -44,7 +44,7 @@ public class DetailsControl : MonoBehaviour {
 			0f,
 			foodManager.GetFoodDescription(),
 			null,
-			(float)foodManager.GetFoodPrice(),
+			foodManager.GetFoodPrice(),
 			1,
 			"",
 			comments
